Conserto rodada 1 da Task 5: bigint→integer, coerção de vehicle_id, teste da trava e casos de mesclar
- Migration: vehicle_id vira integer (era bigint). O driver pg faz parsing
  de int8 pra STRING, não number — apesar do tipo Kysely dizer `number`.
  Map.set com chave string nunca bateria com o Number(v.id) do consumo
  (Task 7), e a sobrescrita humana ficaria protegida no banco e invisível
  pra sempre, sem erro nenhum. Alinha com vehicle_embeddings,
  vehicle_hero_asset e vehicle_section_content, que sempre foram integer.
  Coluna já foi alterada em produção (tabela vazia); esta migration é só
  pra quem aplicar do zero não recriar o problema.
- lerRotulos: Number(l.vehicle_id) defensivo ao montar o Map, comentado,
  pra sobreviver se o tipo da coluna mudar de novo.
- mesclar: troca `if (gravado?.sobrescritoPor)` por `!= null` — string
  vazia não é ausência de sobrescrita.
- db-smoke.test.ts: novo teste compila o upsert de vehicle_semantic_labels
  sem banco e afirma que o WHERE que protege sobrescrito_por está no SQL
  gerado. Verificado que ele quebra se o `.where(...)` sumir do onConflict
  (removi manualmente, teste falhou, restaurei).
- repositorio-rotulos.test.ts: mais três casos de mesclar — arrays vazios
  sob sobrescrita, prosa nula sob sobrescrita, sobrescritoPor === ''.

+++ supabase/migrations/20260816_vehicle_semantic_labels.sql
@@ -4,7 +4,13 @@
 -- regerada do zero a qualquer momento; esta contém correção humana, que não
 -- pode ser perdida numa ressincronização.
 create table if not exists vehicle_semantic_labels (
-  vehicle_id        bigint primary key,
+  -- integer, não bigint: alinhado com vehicle_embeddings, vehicle_hero_asset
+  -- e vehicle_section_content. bigint (int8) faz o driver `pg` devolver
+  -- vehicle_id como STRING, não number — apesar de o tipo do Kysely dizer
+  -- `number` — e isso quebra silenciosamente qualquer Map/comparação por
+  -- id feita rio abaixo (ex.: `gravados.get(Number(v.id))` nunca bate com
+  -- uma chave string).
+  vehicle_id        integer primary key,
   rotulos_uso       text[] not null default '{}',
   rotulos_comprador text[] not null default '{}',
   rotulos_forca     text[] not null default '{}',
